메인 메뉴 분기 처리
- 모바일일 경우 메인 메뉴가 사이드바로 대체
- 검색 구현 하였으나 추후 이동 시킬 예정
- 사이드바로 상단에서 슬라이드 하도록 수정할 예정
